+++ packages/engine/package.json
@@ -1,11 +1,14 @@
 {
-  "name": "@iq-firebolt/server",
+  "name": "@iq-firebolt/engine",
   "version": "0.6.0",
   "description": "> TODO: description",
   "author": "iq team",
   "homepage": "https://github.com/IQ-tech/firebolt#readme",
   "license": "Apache-2.0",
-  "main": "lib/server.js",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "sideEffects": false,
   "directories": {
     "lib": "lib",
     "test": "__tests__"
