+++ packages/engine/src/helpers/validateStep.ts
