--- packages/server/src/design.ts
@@ -1,30 +1,0 @@
-import JSONSample from "./json-sample.json"
-
-
-const myFormExperience = createFireboltEngine({
-  resolvers: {
-    getJSONSchema: JSONSample,
-    getSession: (sessionId: string) => {
-      // asdf
-    },
-    setSession: (sessionId: string) => {
-
-    }
-  },
-  hooks: {
-    
-  }
-})
-
-// ----------------------------------
-
-// uso com express
-import expressFireboltAdapter from "asdfdsad"
-const expressFirebolt = expressFireboltAdapter(myFormExperience)
-
-app.use("/form", expressFirebolt)
-
-
-// uso com serverless
-
-// uso com browser
